+++ src/main/java/io/codelex/advancedTest/cards/CardsTest.java
@@ -1,16 +1,18 @@
 package io.codelex.advancedTest.cards;
 
+import java.math.BigDecimal;
+
 public class CardsTest {
 
   public static void main(String[] args) throws NotEnoughFundsException {
-    CreditCard creditCard = new CreditCard(1848454, "John Smith", "963", 2000.0);
-    DebitCard debitCard = new DebitCard(4545454, "Smith John", "456", 9999);
+    CreditCard creditCard = new CreditCard(1848454, "John Smith", "963", BigDecimal.valueOf(2000));
+    DebitCard debitCard = new DebitCard(4545454, "Smith John", "456", BigDecimal.valueOf(9999));
     System.out.println(creditCard.getBalance());
     System.out.println(debitCard.getBalance());
-    creditCard.takeMoney(1901.15);
-    debitCard.addMoney(2.5);
+    creditCard.takeMoney(BigDecimal.valueOf(1901.15));
+    debitCard.addMoney(BigDecimal.valueOf(2.5));
     System.out.println(creditCard.getBalance());
     System.out.println(debitCard.getBalance());
-    creditCard.takeMoney(100);
+    creditCard.takeMoney(BigDecimal.valueOf(100));
   }
 }
